% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_context_measures.R
\name{closest_context_measures}
\alias{closest_context_measures}
\title{Find close Cancer InFocus measure matches}
\usage{
closest_context_measures(measure, available_measures, n = 5)
}
\arguments{
\item{measure}{Unsupported measure name.}

\item{available_measures}{A data frame returned by \code{\link[=available_context_measures]{available_context_measures()}}.}

\item{n}{Maximum number of close matches to return.}
}
\value{
A character vector of suggested measure definitions.
}
\description{
Finds likely valid Cancer InFocus measure definitions for an unsupported
measure name.
}
\keyword{internal}
