% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_cache_path.R
\name{context_cache_path}
\alias{context_cache_path}
\title{Create a cache path for Cancer InFocus context files}
\usage{
context_cache_path(url)
}
\arguments{
\item{url}{URL for a hosted ContextLinkr data file.}
}
\value{
A character string containing the local cache path.
}
\description{
Creates a stable local cache path for a hosted Cancer InFocus ContextLinkr
data URL.
}
\keyword{internal}
