% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_filters.R
\name{context_failures}
\alias{context_failures}
\title{Filter records not successfully joined to contextual data}
\usage{
context_failures(.data)
}
\arguments{
\item{.data}{A data frame returned by \code{\link[=join_context]{join_context()}} or another
ContextLinkr workflow containing \code{.context_joined}.}
}
\value{
A tibble containing rows where \code{.context_joined} is not \code{TRUE}.
}
\description{
\code{context_failures()} returns rows that were not successfully joined to
contextual data. For results from \code{\link[=join_context]{join_context()}}, this means
\code{.context_joined} is \code{FALSE} or missing.
}
\examples{
joined <- tibble::tibble(
  id = 1:3,
  .context_joined = c(TRUE, FALSE, NA)
)

context_failures(joined)

}
\seealso{
\code{\link[=join_context]{join_context()}}, \code{\link[=context_summary]{context_summary()}}, \code{\link[=link_summary]{link_summary()}},
\code{\link[=missing_context_keys]{missing_context_keys()}}, \code{\link[=context_successes]{context_successes()}}
}
