% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_provenance.R
\name{context_provenance}
\alias{context_provenance}
\title{Report ContextLinkr context provenance}
\usage{
context_provenance(x)
}
\arguments{
\item{x}{A ContextLinkr context output, such as a result from \code{\link[=get_context]{get_context()}},
\code{\link[=add_context]{add_context()}}, or \code{\link[=link_context]{link_context()}} with \code{include_context = TRUE}.}
}
\value{
A tibble containing available provenance metadata.
}
\description{
\code{context_provenance()} reports provenance metadata attached to ContextLinkr
context outputs.
}
\examples{
\dontrun{
context <- get_context(
  geographies = "11001006202",
  measures = "Total Population",
  geography = "tract"
)

context_provenance(context)
}

}
