% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_filters.R
\name{context_successes}
\alias{context_successes}
\title{Filter records successfully joined to contextual data}
\usage{
context_successes(.data)
}
\arguments{
\item{.data}{A data frame returned by \code{\link[=join_context]{join_context()}} or another
ContextLinkr workflow containing \code{.context_joined}.}
}
\value{
A tibble containing rows where \code{.context_joined} is \code{TRUE}.
}
\description{
\code{context_successes()} returns rows that were successfully joined to
contextual data. For results from \code{\link[=join_context]{join_context()}}, this means
\code{.context_joined} is \code{TRUE}.
}
\examples{
joined <- tibble::tibble(
  id = 1:3,
  .context_joined = c(TRUE, FALSE, TRUE)
)

context_successes(joined)

}
\seealso{
\code{\link[=join_context]{join_context()}}, \code{\link[=link_summary]{link_summary()}}, \code{\link[=context_failures]{context_failures()}}
}
