% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextLinkr-package.R
\docType{package}
\name{ContextLinkr-package}
\alias{ContextLinkr}
\alias{ContextLinkr-package}
\title{ContextLinkr: Link Individual Records to Geographic Context}
\description{
ContextLinkr provides tools for linking individual-level records to
geographic context. The package supports modular workflows for geocoding
address data, identifying Census tracts from coordinates, linking records to
contextual datasets, and reviewing workflow quality.
}
\section{Geocoding}{

Use \code{\link[=gc_address]{gc_address()}} to geocode address data when appropriate. Because address
data can be sensitive, external geocoding requires explicit confirmation with
\code{confirm_external = TRUE}.

Use \code{\link[=geocode_summary]{geocode_summary()}}, \code{\link[=geocode_successes]{geocode_successes()}}, and \code{\link[=geocode_failures]{geocode_failures()}} to
review geocoding results.
}

\section{Tract identification}{

Use \code{\link[=id_tract]{id_tract()}} to identify Census tracts from latitude and longitude.
This workflow can be used after geocoding or directly with pre-geocoded data.

Use \code{\link[=tract_summary]{tract_summary()}}, \code{\link[=tract_successes]{tract_successes()}}, and \code{\link[=tract_failures]{tract_failures()}} to review
tract identification results.
}

\section{End-to-end linking}{

Use \code{\link[=link_context]{link_context()}} as a convenience wrapper for geocoding records when
needed and identifying Census tracts.

Use \code{\link[=link_summary]{link_summary()}}, \code{\link[=link_successes]{link_successes()}}, and \code{\link[=link_failures]{link_failures()}} to review
linked outputs.
}

\section{Contextual data joins}{

Use \code{\link[=join_context]{join_context()}} to join linked records to contextual datasets that are
already available in memory. Use \code{\link[=missing_context_keys]{missing_context_keys()}} to identify linked
geographic keys that are not present in a contextual dataset.

Use \code{\link[=context_summary]{context_summary()}}, \code{\link[=context_successes]{context_successes()}}, and \code{\link[=context_failures]{context_failures()}} to
review contextual data joins.
}

\section{Example data}{

The package includes \link{sample_addresses} and \link{sample_context} as small
illustrative datasets for examples and tests. These datasets are not
intended for analysis.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CIOData/ContextLinkr}
  \item Report bugs at \url{https://github.com/CIOData/ContextLinkr/issues}
}

}
\author{
\strong{Maintainer}: Todd Burus \email{tburus@uky.edu}

Authors:
\itemize{
  \item Todd Burus \email{tburus@uky.edu}
}

}
\keyword{internal}
