% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_context_measure_suggestions.R
\name{format_context_measure_suggestions}
\alias{format_context_measure_suggestions}
\title{Format Cancer InFocus measure suggestions}
\usage{
format_context_measure_suggestions(suggestions)
}
\arguments{
\item{suggestions}{Character vector of suggested measure definitions.}
}
\value{
A single character string.
}
\description{
Formats close Cancer InFocus measure matches for inclusion in an error
message.
}
\keyword{internal}
