% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_successes.R
\name{geocode_successes}
\alias{geocode_successes}
\title{Extract successful geocoding results}
\usage{
geocode_successes(x)
}
\arguments{
\item{x}{A data frame returned by \code{\link[=gc_address]{gc_address()}}.}
}
\value{
A tibble containing rows where \code{.geocoded} is \code{TRUE}.
}
\description{
Returns records that were successfully geocoded by \code{\link[=gc_address]{gc_address()}}.
}
\examples{
x <- tibble::tibble(
  id = 1:2,
  latitude = c(38.9, NA_real_),
  longitude = c(-77.0, NA_real_),
  .geocoded = c(TRUE, FALSE),
  .geocode_input = c("components", "components")
)

geocode_successes(x)

}
