% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_summary.R
\name{geocode_summary}
\alias{geocode_summary}
\title{Summarize geocoding results}
\usage{
geocode_summary(x)
}
\arguments{
\item{x}{A data frame returned by \code{\link[=gc_address]{gc_address()}}.}
}
\value{
A tibble with geocoding summary information.
}
\description{
Returns a compact summary of geocoding results created by \code{\link[=gc_address]{gc_address()}}.
}
\examples{
x <- sample_addresses
attr(x, "contextlinkr_geocode_summary") <- list(
  matched = 5,
  total = 5,
  match_rate = 1,
  geocoder = "osm"
)

geocode_summary(x)

}
