% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_context.R
\name{get_context}
\alias{get_context}
\title{Retrieve Cancer InFocus contextual data}
\usage{
get_context(geographies, measures = NULL, geography = "tract", year = NULL)
}
\arguments{
\item{geographies}{A character vector of geographic identifiers, such as
Census tract GEOIDs.}

\item{measures}{Optional character vector of contextual measures to retrieve.
If \code{NULL}, a default set of measures may be returned in future versions.}

\item{geography}{Geographic level. Currently planned values include
\code{"tract"} and \code{"county"}.}

\item{year}{Optional year or data vintage.}
}
\value{
Currently aborts with an informative message because Cancer InFocus
retrieval is not yet implemented.
}
\description{
\code{get_context()} will retrieve contextual variable values from Cancer InFocus
data for specified geographic identifiers.
}
\details{
This function is planned but not yet implemented. Future versions will use
Cancer InFocus data as the primary source of contextual variables for
individual-level linkage workflows.
}
\examples{
\dontrun{
get_context(
  geographies = c("11001980000", "24510040100"),
  measures = c("poverty", "rurality"),
  geography = "tract"
)
}

}
