% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_remote_path.R
\name{is_remote_path}
\alias{is_remote_path}
\title{Detect remote paths}
\usage{
is_remote_path(path)
}
\arguments{
\item{path}{Local path or URL.}
}
\value{
A logical value.
}
\description{
Detects whether a path is an HTTP or HTTPS URL.
}
\keyword{internal}
