% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_filters.R
\name{link_failures}
\alias{link_failures}
\title{Filter records that were not successfully linked}
\usage{
link_failures(.data)
}
\arguments{
\item{.data}{A data frame returned by \code{\link[=link_context]{link_context()}} or another
ContextLinkr workflow containing \code{.tract_identified}.}
}
\value{
A tibble containing rows where \code{.tract_identified} is not \code{TRUE}.
}
\description{
\code{link_failures()} returns rows that were not successfully linked to Census
tract geography. For results from \code{\link[=link_context]{link_context()}}, this means
\code{.tract_identified} is \code{FALSE} or missing.
}
\examples{
linked <- tibble::tibble(
  id = 1:3,
  .tract_identified = c(TRUE, FALSE, NA)
)

link_failures(linked)

}
\seealso{
\code{\link[=link_context]{link_context()}}, \code{\link[=link_summary]{link_summary()}}, \code{\link[=link_successes]{link_successes()}}
}
