% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_filters.R
\name{link_successes}
\alias{link_successes}
\title{Filter successfully linked records}
\usage{
link_successes(.data)
}
\arguments{
\item{.data}{A data frame returned by \code{\link[=link_context]{link_context()}} or another
ContextLinkr workflow containing \code{.tract_identified}.}
}
\value{
A tibble containing rows where \code{.tract_identified} is \code{TRUE}.
}
\description{
\code{link_successes()} returns rows that were successfully linked to Census
tract geography. For results from \code{\link[=link_context]{link_context()}}, this means
\code{.tract_identified} is \code{TRUE}.
}
\examples{
linked <- tibble::tibble(
  id = 1:3,
  .tract_identified = c(TRUE, FALSE, TRUE)
)

link_successes(linked)

}
\seealso{
\code{\link[=link_context]{link_context()}}, \code{\link[=link_summary]{link_summary()}}, \code{\link[=link_failures]{link_failures()}}
}
