% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract_failures.R
\name{tract_failures}
\alias{tract_failures}
\title{Extract failed Census tract identification results}
\usage{
tract_failures(x)
}
\arguments{
\item{x}{A data frame returned by \code{\link[=id_tract]{id_tract()}}.}
}
\value{
A tibble containing rows where \code{.tract_identified} is \code{FALSE}.
}
\description{
Returns records that were not assigned to Census tracts by \code{\link[=id_tract]{id_tract()}}.
}
\examples{
x <- tibble::tibble(
  id = 1:3,
  tract_geoid = c("11001980000", NA_character_, "11001006202"),
  .tract_identified = c(TRUE, FALSE, TRUE),
  .tract_state_fips = c("11", "11", "11"),
  .tract_year = c(2020, 2020, 2020)
)

tract_failures(x)

}
